Set preconditioner always
Some tests (Mainly which use UMFPACK) seem to have a dependency on
the Ifpack preconditioner and requires the preconditioner to be set
(which needs to be removed). Thus the Ifpack preconditioner is structured
in a way that it can also initialize the preconditioner when no Ifpack
xml file is specified in the input files.